Organize transportation pkg, add relationships, sync schema, tidy docs
This commit organizes the transportation Pydantic package to align with
the other theme packages already done: addresses, base, buildings,
divisions, and places. Transportation was the last theme still using the
old layout, so with this commit the whole theme set is organized
consistently.

Organization work done:

1. Each feature type gets its own main module. `connector` becomes a
   flat `connector.py`, but because `segment` is a discriminated union
   over road, rail, and water subtypes, `segment` becomes a *package*
   with one module per subtype (`road.py`, `rail.py`, `water.py`) plus a
   private `_common.py` for the shared base. This is the one place we do
   NOT eliminate a sub-directory (unlike divisions): the subtype union
   makes a package the natural fit.
2. Types used exclusively by one subtype move into that subtype's module
   and are renamed to be subtype-scoped: `Subtype` -> `SegmentSubtype`,
   `Subclass` -> `RoadSubclass`, `SubclassRule` -> `RoadSubclassRule`,
   `SurfaceRule` -> `RoadSurfaceRule`, `AccessRestrictionRule` ->
   `AccessRule`, `SequenceEntry` -> `ProhibitedTransitionSequenceEntry`.
   Exception (deliberate): the mode-independent rule types currently used
   only by road but conceivably reusable by other subtypes (the
   prohibited-transition, speed, and width types) are kept in `_common.py`.
3. Eliminated the old flat `enums.py`, `models.py`, and `types.py`, and
   the `connector/` and `segment/` (models.py) sub-packages.
4. Curated `__all__` in both the theme `__init__.py` and the segment
   package's `__init__.py`.

Relationships added:

1. `from_connector_id` and `to_connector_id` in `DestinationRule` (to
   `Connector`).
2. `to_segment_id` in `DestinationRule` (to `TransportationSegment`).
3. `connector_id` in `ProhibitedTransitionSequenceEntry` (to `Connector`).
4. `segment_id` in `ProhibitedTransitionSequenceEntry` (to
   `TransportationSegment`).

Docs tidied and updated:

1. Eliminated all `make docformat` errors pertaining to transportation.
2. Converted `SegmentSubtype` and `AccessType` to `DocumentedEnum` so
   they have a richer presence in the schema reference. (`RoadSurface`
   and `DestinationLabelType` are intentionally left as plain enums for
   now, as I don't have cycles to generate docs for them.)

Schema reconciliation (a dimension the divisions refactor did not have):

1. Made `routes` and `subclass_rules` road-only (moved from the shared
   base to `RoadSegment`) and back-ported the change to the JSON Schema
   so both representations agree. While technically backwardly-
   incompatible, this fixes a error that dates back to the JSON Schema
   and that was never intended.
2. Removed the scalar `level` from segment in both the Pydantic model and
   the JSON Schema, finishing the model/schema sync begun in "Remove
   levels from types without data for them" (commit dd5d3f2a, PR #390),
   which updated the model but left `level` on the legacy JSON Schema,
   leading to a discrepancy between schemas.
