+++ examples/transportation/segment/road/road.yaml
@@ -28,7 +28,6 @@ properties:
     - value: gravel
   road_flags:
     - values: [is_link, is_tunnel] # Note: `is_link` is deprecated and will be removed in a future release in favor of the link subclass
-  level: -1
   level_rules:
     - value: -1
   width_rules:
