+++ packages/overture-schema-divisions-theme/tests/division_boundary_baseline_schema.json
@@ -564,6 +564,6 @@
     "geometry",
     "properties"
   ],
-  "title": "boundary",
+  "title": "division_boundary",
   "type": "object"
 }
