--- packages/overture-schema-transportation-theme/src/overture/schema/transportation/connector/__init__.py
@@ -1,3 +1,0 @@
-from .models import Connector
-
-__all__ = ["Connector"]
